requirements.txt
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-text-splitters

transformers
huggingface_hub
sentence-transformers

langchain-ollama

pinecone-client
langchain-pinecone

streamlit
flask
pypdf
python-dotenv
pinecone[grpc]
